+++ src/core/components/SkDateTimePicker/SkDateTimePicker.constants.ts
@@ -1,0 +1,2 @@
+export const formatDate = 'yyyy-MM-dd';
+export const formatTime = 'H:mm';
