+++ src/core/components/SkDateTimePicker/SkDateTimeRangeFilter.enum.ts
@@ -1,0 +1,6 @@
+export enum SkDateTimePickerLabels {
+  CalendarTitlePicker = 'Absolute time range',
+  StarDatePicker = 'From',
+  EndDatePicker = 'To',
+  CalendarPickerButton = 'Apply time range'
+}
