+++ src/core/components/SkSearchFilter/SkSearchFilter.enum.ts
